arm64: head.S: Convert to modern annotations for assembly functions
In an effort to clarify and simplify the annotation of assembly functions
in the kernel new macros have been introduced. These replace ENTRY and
ENDPROC and also add a new annotation for static functions which previously
had no ENTRY equivalent. Update the annotations in the core kernel code to
the new macros.

Signed-off-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Change-Id: Icea9c2f01d58f67885d02a02d8241ab9f65069e6
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>