vendor_hook: sched: Add Adaptive CPU procfs file
The Adaptive CPU model in userspace needs stats about CPU frequency
residencies and idle periods. While the best approach to get those out
of the kernel is to use an eBPF program, this faces unrelated but
non-trivial issues to load an eBPF program from vendor code (see
b/162057235).

This attempts to work-around the issue by re-implementing ag/14424500
using a proper kernel module, for which permissions are easier to
manage. The module exposes a new binary file in procfs in the vendor
sched folder: acpu_stats. This file holds the content of a per-CPU
array with relevant stats in binary format.

Bug: 207662659
Change-Id: I3cea3a90ff656fb9a9f6baef6527e792ed3a5dcf
Signed-off-by: Quentin Perret <qperret@google.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>