sched: suppress silver region frequency boost
Currently, we boost frequencies by 25% across the board. However, due to
region 3 silver CPUs energy concerns, we don't want to apply this boost
provided that the utilization on silver CPUs is high enough.
Additionally, there is the caveat that if the contribution from RT task
is high enough (currently hard coded to 25% of silver capacity), we
continue to apply the boost.

To manipulate the silver threshold (set to 1000 by default, therefore
this is disabled), use:

echo x > /proc/sys/kernel/sched_silver_thres .

Change-Id: If689d0935f0c60f7ccc19e63780286c815eb801a
Signed-off-by: Shaleen Agrawal <shalagra@codeaurora.org>
Signed-off-by: DennySPb <dennyspb@gmail.com>