sched/walt: Fix lock_cpus mask setting during IC migrations
In the existing code, we set the notifier pending flag only on the
source and destination CPU. But, in irq_work_restrict_to_mig_clusters(),
it clears the lock_cpus mask upon visiting non-involved CPUs in the
cluster. This ends up not locking any CPUs, essentially not updating CPU
frequencies upon inter cluster migrations.

Fix this by only clearly lock_cpus if none of the CPUs in the cluster
are involved in a migration.

Change-Id: I486e6c62bf306de8fddbd3eab0537c419aedaff2
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>