i2c-msm-geni: KASAN: use-after-free in __list_add_valid+0x2c/0xc4
This UAF issue is seen when driver is removed and inserted.

During driver removal, pm runtime resume callback invoked
in which as part of clock, ab/ib nodes are added in
common struct geni_se_dev. As part of driver exit,
we are not removing the ab/ib list from common structure list
due to which the issue is seen when driver is loaded.

As part of driver removal, checking the status of runtime suspend
if it is not suspended, invoke geni suspend call otherwise ignore.
So by suspend call ensured that ab/ib are removed from lists,
so that UAF will not be encountered when next load of driver.

Change-Id: I1f0c7a29c5e268a1ab5c017e271ad0484dcab24f
Signed-off-by: Praveen Talari <quic_ptalari@quicinc.com>
Signed-off-by: Vaishnavi AVS <quic_vavs@quicinc.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>